#146 refactor: split webapp.rs into logical modules
Split monolithic webapp.rs (3591 lines) into 10 focused modules following Single Responsibility Principle:

- types.rs (265 lines): All shared types, enums, and structs
- core.rs (153 lines): Instance management and validation
- lifecycle.rs (235 lines): App lifecycle and window control
- buttons.rs (701 lines): All button operations (main, secondary, bottom, back)
- dialogs.rs (102 lines): User dialogs and popups
- navigation.rs (232 lines): Links, sharing, and navigation
- permissions.rs (196 lines): Permission requests and file operations
- theme.rs (53 lines): Theme and color management
- viewport.rs (92 lines): Viewport and safe area operations
- events.rs (276 lines): Event handling system

Main webapp.rs reduced to 1401 lines (re-exports + tests).

Benefits:
- Each module has single clear responsibility
- Easier maintenance and navigation
- Better encapsulation with pub(super) visibility
- No module exceeds 701 lines
- All tests pass (207 tests)
- Zero clippy warnings
